% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustSegments.R
\name{AdjustSegments}
\alias{AdjustSegments}
\title{Adjust (merged) segment information}
\usage{
AdjustSegments(si.input, mp, ms)
}
\arguments{
\item{si.input}{(dataFrame): DataFrame containing the SSI info.}

\item{mp}{(character): Name of the primary marker.}

\item{ms}{(character): Name of the secondary marker.}
}
\value{
(dataFrame) Returns a dataframe with an update SSI information dataFrame.
}
\description{
Adjust (merged) segment information
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "CompensAID"))

# Marker names
primary.marker <- "CD19"
secondary.marker <- "CD3"

# Channel names
cp <- "PE-Cy5-A"
cs <- "PE-Cy7-A"

# Density-based cut-off detection
center.plot <- 2
co <- DensityGating(og = file,
                    cp.value = center.plot)

# Get combinations
mc <- GetMarkerCombinations(og = file)

# Parameter for the distance between the primary positive and negative population
separation.distance <- 0.25

# Obtain empty SSI matrix information
range.value <- 4
si <- EmptyMatrixInfo(og = file,
                      rv.input = range.value,
                      mc.input = mc,
                      co.input = co,
                      sd.input = separation.distance)

# Obtain populations
pop <- GetPopulations(og = file,
                      primary = primary.marker,
                      secondary = secondary.marker,
                      co.input = co,
                      sd.input = separation.distance)

# Select primary positive population
population.positive <- pop$primary.positive

# Get the events that fall within the fourth segment
segment.value <- 4
range.value <- (max(population.positive[, cp]) - min(population.positive[, cp]))/segment.value

# Adjust the SSI info if there are >20 events in the (segmented) positive population
# (output = "PASS"); meaning more events present than threshold
si <- UpdateMatrixInfo(si.input = si,
                       rv.input = segment.value,
                       range.input = range.value,
                       primary = primary.marker,
                       secondary = secondary.marker,
                       output = "PASS",
                       population = pop)

# Update segments; function calls AdjustSegments
events.value <- 20
si <- UpdateSegments(si.input = si,
                     primary = primary.marker,
                     secondary = secondary.marker,
                     ev.input = events.value,
                     rv.input = segment.value)

}
\seealso{
\code{\link{CompensAID}}, \code{\link{DensityGating}}, \code{\link{EmptyMatrixInfo}}, \code{\link{UpdateMatrixInfo}}, \code{\link{GetPopulations}}
}
