% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSSI.R
\name{CalculateSSI}
\alias{CalculateSSI}
\title{Calculate Secondary Stain Index}
\usage{
CalculateSSI(si.input, primary.channel, secondary.channel, segment)
}
\arguments{
\item{si.input}{(dataFrame): DataFrame containing the SSI info.}

\item{primary.channel}{(character): Name of the primary channel.}

\item{secondary.channel}{(character): Name of the secondary channel.}

\item{segment}{(numerical): Numerical value determining for which segment the SSI should be calculated.}
}
\value{
(numerical) Returns the Secondary Stain Index score.
}
\description{
Calculate Secondary Stain Index
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "compensAID"))

# Parameter for the number of segments
range.value <- 4

# Density-based cut-off detection
center.plot <- 2
co <- DensityGating(og = file,
                    cp.value = center.plot)

# Get combinations
mc <- GetMarkerCombinations(og = file)

# Parameter for the distance between the primary positve and negative population
separation.distance <- 0.25

# Obtain SSI info dataFrame
si <- EmptyMatrixInfo(og = file,
                      rv.input = range.value,
                      mc.input = mc,
                      co.input = co,
                      sd.input = separation.distance)

# Channel names
cp <- "PE-Cy5-A"
cs <- "PE-Cy7-A"

# Calculate SSI for the last segment
CalculateSSI(si.input = si,
             primary.channel = cp,
             secondary.channel = cs,
             segment = range.value)

}
\seealso{
\code{\link{CompensAID}}, \code{\link{EmptyMatrixInfo}}, \code{\link{DensityGating}}
}
