% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompensAID.R
\name{CompensAID}
\alias{CompensAID}
\title{Run compensAID tool; automated detection of compensation issues}
\usage{
CompensAID(
  ff,
  range.value = 4,
  events.value = 20,
  center.plot = 2,
  separation.distance = 0.25
)
}
\arguments{
\item{ff}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}

\item{range.value}{(numerical): Numerical value of the number of segments in the primary positive population.}

\item{events.value}{(numerical): Numerical value of the minimum number of events per population/segment.}

\item{center.plot}{(numerical): Numerical value determining the preliminary center.}

\item{separation.distance}{(numerical): Numerical value determining the distance between the primary negative and positive population.}
}
\value{
(list) Returns a list containing the full SSI output and SSI matrix
}
\description{
Run compensAID tool; automated detection of compensation issues
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "compensAID"))

# Run compensAID tool
segments <- 4
min.events <- 20
center <- 2
distance.populations <- 0.25
compensAID.res <- CompensAID(ff = file,
                             range.value = segments,
                             events.value = min.events,
                             center.plot = center,
                             separation.distance = distance.populations)

}
