% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityGating.R
\name{DensityGating}
\alias{DensityGating}
\title{Perform density-based cut-off detection}
\usage{
DensityGating(og, cp.value)
}
\arguments{
\item{og}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}

\item{cp.value}{(numerical): Numerical value determining the preliminary center.}
}
\value{
(dataFrame) Returns a dataFrame with the output of the density-based cut-off detection
}
\description{
Perform density-based cut-off detection
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "compensAID"))

# Perform density-based cut-off detection
center.plot <- 2
co <- DensityGating(og = file,
                    cp.value = center.plot)

}
\seealso{
\code{\link{CompensAID}, \link{GetClosestCenter}}
}
