% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustSegments.R
\name{.AdjustSegments}
\alias{.AdjustSegments}
\title{Adjust (merged) segment information}
\usage{
.AdjustSegments(si.input, mp, ms)
}
\arguments{
\item{si.input}{(dataFrame): DataFrame containing the SSI info.}

\item{mp}{(character): Name of the primary marker.}

\item{ms}{(character): Name of the secondary marker.}
}
\value{
Dataframe with updated SSI info
}
\description{
Adjust (merged) segment information
}
\keyword{internal}
