% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSSI.R
\name{.CalculateSSI}
\alias{.CalculateSSI}
\title{Calculate the Secondary Stain Index for each marker combination}
\usage{
.CalculateSSI(si.input, primary.channel, secondary.channel, segment)
}
\arguments{
\item{si.input}{(dataFrame): DataFrame containing the SSI info.}

\item{primary.channel}{(character): Name of the primary marker.}

\item{secondary.channel}{(character): Name of the secondary marker.}

\item{segment}{(numeric): Numerical value of the segment}
}
\value{
Numeric value of Secondary Stain Index score
}
\description{
Calculate the Secondary Stain Index for each marker combination
}
\keyword{internal}
