% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanSegments.R
\name{.CleanSegments}
\alias{.CleanSegments}
\title{Clean adjust segment information}
\usage{
.CleanSegments(si.input, mp, ms)
}
\arguments{
\item{si.input}{(dataFrame): DataFrame containing the SSI info.}

\item{mp}{(character): Name of the primary marker.}

\item{ms}{(character): Name of the secondary marker.}
}
\value{
(dataFrame) Returns a dataframe with an update SSI information dataFrame.
}
\description{
Clean adjust segment information
}
\keyword{internal}
