% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetermineCenter.R
\name{.DetermineCenter}
\alias{.DetermineCenter}
\title{Rougly estimate the 'center' of a marker}
\usage{
.DetermineCenter(og)
}
\arguments{
\item{og}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}
}
\value{
(list) Returns a list containing the estimated center and additional margin
}
\description{
Rougly estimate the 'center' of a marker
}
\keyword{internal}
