% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyMatrixInfo.R
\name{.EmptyMatrixInfo}
\alias{.EmptyMatrixInfo}
\title{Obtain empty SSI dataFrame}
\usage{
.EmptyMatrixInfo(
  og = ff,
  rv.input = range.value,
  mc.input = mc,
  co.input = co,
  sd.input = separation.distance
)
}
\arguments{
\item{og}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}

\item{rv.input}{(numerical): Numerical value of the number of segments in the primary positive population.}

\item{mc.input}{(dataFrame): dataFrame containing all possible marker combinations.}

\item{co.input}{(dataFrame) dataFrame with the output of the density-based cut-off detection.}

\item{sd.input}{(numerical): Numerical value determining the distance between the primary negative and positive population.}
}
\value{
(dataFrame) Returns an empty SSI dataFrame
}
\description{
Obtain empty SSI dataFrame
}
\keyword{internal}
