% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventRequirement.R
\name{.EventRequirement}
\alias{.EventRequirement}
\title{Assess event requirement}
\usage{
.EventRequirement(negative, positive, events.value)
}
\arguments{
\item{negative}{(matrix): All events that fall within the negative primary population.}

\item{positive}{(matrix): All events that fall within the positive primary population.}

\item{events.value}{(numerical): Numerical value defining the minimum requirement of events.}
}
\value{
(boolean) Returns TRUE/FALSE depending on the required number of events.
}
\description{
Assess event requirement
}
\keyword{internal}
