% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClosestCenter.R
\name{.GetClosestCenter}
\alias{.GetClosestCenter}
\title{Identify value closest to visual estimation}
\usage{
.GetClosestCenter(row, closest)
}
\arguments{
\item{row}{(numerical): Numerical values of all density-based cut-off detection values.}

\item{closest}{(numerical): Numerical value determining the preliminary center.}
}
\value{
(numerical) Returns the value closest to the visual center of the plot.
}
\description{
Identify value closest to visual estimation
}
\keyword{internal}
