% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClosestLimit.R
\name{.GetClosestLimit}
\alias{.GetClosestLimit}
\title{Detect if smaller peaks inclusion improved the density-based cut-off detection.}
\usage{
.GetClosestLimit(old.limit, new.limit, center.plot)
}
\arguments{
\item{old.limit}{(numerical): Numerical value of the limit detected under default settings.}

\item{new.limit}{(numerical): Numerical value of the limit detected under adjusted settings}

\item{center.plot}{(numerical): Numerical value of the estimated center.}
}
\value{
(numerical) Returns the limit that fits the data best.
}
\description{
Detect if smaller peaks inclusion improved the density-based cut-off detection.
}
\keyword{internal}
