% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulations.R
\name{.GetPopulations}
\alias{.GetPopulations}
\title{Obtain (segmented) populations}
\usage{
.GetPopulations(og, primary, secondary, co.input, sd.input)
}
\arguments{
\item{og}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}

\item{primary}{(character): Name of the primary marker.}

\item{secondary}{(character): Name of the secondary marker.}

\item{co.input}{(dataFrame) dataFrame with the output of the density-based cut-off detection.}

\item{sd.input}{(numerical): Numerical value determining the distance between the primary negative and positive population.}
}
\value{
(list) Returns a list with the negative population of the primary and secondary marker, and positive population of the primary marker.
}
\description{
Obtain (segmented) populations
}
\keyword{internal}
