% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSegment.R
\name{.GetSegment}
\alias{.GetSegment}
\title{Calculate segment information}
\usage{
.GetSegment(
  population,
  primary.channel,
  secondary.channel,
  segment,
  range.input
)
}
\arguments{
\item{population}{(matrix): Matrix of the population that is segmented.}

\item{primary.channel}{(character): Name of the primary channel.}

\item{secondary.channel}{(character): Name of the secondary channel.}

\item{segment}{(numerical) Segment number for which the information is obtained.}

\item{range.input}{(numerical): Numerical value defining the width of each segment.}
}
\description{
Calculate segment information
}
\keyword{internal}
