% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeSegments.R
\name{.MergeSegments}
\alias{.MergeSegments}
\title{Combine segments that do not meet the requirements}
\usage{
.MergeSegments(segment.event.count, ev.input)
}
\arguments{
\item{segment.event.count}{(numerical): Number of events per segment}

\item{ev.input}{(numerical): Minimum required number of events.}
}
\value{
(list) Returns a list with which segments need to be merged.
}
\description{
Combine segments that do not meet the requirements
}
\keyword{internal}
