% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateMatrixInfo.R
\name{.UpdateMatrixInfo}
\alias{.UpdateMatrixInfo}
\title{Calculate segment information}
\usage{
.UpdateMatrixInfo(
  si.input,
  range.input = NULL,
  primary,
  secondary,
  output,
  rv.input,
  population = NULL
)
}
\arguments{
\item{si.input}{(dataFrame): dataFrame containing SSI info.}

\item{range.input}{(numerical): Width of each segment.}

\item{primary}{(character): Name of the primary marker.}

\item{secondary}{(character) Name of the secondary marker.}

\item{output}{(character): Will be "PASS" or "No positive/negative population" depending on the required events.}

\item{rv.input}{(numerical): Number of segments.}

\item{population}{(matrix): Matrix with the population for which the information is obtained.}
}
\value{
(dataFrame) Returns a dataframe with an update SSI information dataFrame.
}
\description{
Calculate segment information
}
\keyword{internal}
