% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateSegments.R
\name{.UpdateSegments}
\alias{.UpdateSegments}
\title{Combine segments that do not meet the requirements}
\usage{
.UpdateSegments(si.input, primary, secondary, ev.input, rv.input)
}
\arguments{
\item{si.input}{(dataFrame): dataFrame containing SSI info.}

\item{primary}{(character): Name of the primary marker.}

\item{secondary}{(character) Name of the secondary marker.}

\item{ev.input}{(numerical): Minimum required number of events.}

\item{rv.input}{(numerical): Number of segments.}
}
\value{
(dataFrame) Returns a dataframe with an update SSI information dataFrame.
}
\description{
Combine segments that do not meet the requirements
}
\keyword{internal}
