% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateSSI.R
\name{.UpdateSSI}
\alias{.UpdateSSI}
\title{Update Secondary Stain Index}
\usage{
.UpdateSSI(si.input, rv.input, primary, secondary, population)
}
\arguments{
\item{si.input}{(dataFrame): DataFrame containing the SSI info.}

\item{rv.input}{(numerical): Numerical value for the number of segments.}

\item{primary}{(character): Name of the primary marker}

\item{secondary}{(character): Name of the secondary marker}

\item{population}{(list): List with the population matrices}
}
\value{
(numerical) Returns the Secondary Stain Index score.
}
\description{
Update Secondary Stain Index
}
\keyword{internal}
