% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmptyMatrix.R
\name{EmptyMatrix}
\alias{EmptyMatrix}
\title{Obtain empty SSI matrix}
\usage{
EmptyMatrix(og)
}
\arguments{
\item{og}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}
}
\value{
(matrix) Returns an empty SSI matrix
}
\description{
Obtain empty SSI matrix
}
\examples{
# Import FCS file
file <- flowCore::read.FCS("path/to/exampleFCS.fcs")

# Empty matrix
sm <- EmptyMatrix(og = file)

}
\seealso{
\code{\link{CompensAID}}
}
