% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventRequirement.R
\name{EventRequirement}
\alias{EventRequirement}
\title{Assess event requirement}
\usage{
EventRequirement(negative, positive, events.value)
}
\arguments{
\item{negative}{(matrix): All events that fall within the negative primary population.}

\item{positive}{(matrix): All events that fall within the positive primary population.}

\item{events.value}{(numerical): Numerical value defining the minimum requirement of events.}
}
\value{
(boolean) Returns TRUE/FALSE depending on the required number of events.
}
\description{
Assess event requirement
}
\examples{
# Import FCS file
file <- flowCore::read.FCS("path/to/exampleFCS.fcs")

# Marker names
primary.marker <- "CD19"
secondary.marker <- "IgL"

# Obtain density-based cut-off detection
center.plot <- 2
co <- DensityGating(og = file, 
                    cp.value = center.plot)
              
# Obtain populations
separation.distance <- 0.25
pop <- GetPopulations(og = file,
                      primary = primary.marker,
                      secondary = secondary.marker,
                      co.input = co,
                      sd.input = separation.distance)
                      
# Select populations                       
primary.negative <- pop$primary.negative
primary.positive <- pop$primary.positive

# Assess event requirement
events.value <- 20
EventRequirement(primary.negative, primary.positive, events.value))

}
\seealso{
\code{\link{CompensAID}, \link{GetPopulations}, \link{}}
}
