% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClosestCenter.R
\name{GetClosestCenter}
\alias{GetClosestCenter}
\title{Identify value closest to visual estimation}
\usage{
GetClosestCenter(row, closest)
}
\arguments{
\item{row}{(numerical): Numerical values of all density-based cut-off detection values.}

\item{closest}{(numerical): Numerical value determining the preliminary center.}
}
\value{
(numerical) Returns the value closest to the visual center of the plot.
}
\description{
Identify value closest to visual estimation
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "CompensAID"))
primary.channel <- "PE-Cy5-A"

# Density-based cut-off detection of a single channel.
density.detection <- flowDensity::deGate(file,
                                         channel = primary.channel,
                                         verbose = FALSE,
                                         all.cuts = TRUE,
                                         upper = TRUE)

# Identify cut-off closest to visual estimation
center.plot <- 2
cut.off <- GetClosestCenter(density.detection,
                            closest = center.plot)

}
\seealso{
\code{\link{CompensAID}, \link{DensityGating}}
}
