% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarkerCombinations.R
\name{GetMarkerCombinations}
\alias{GetMarkerCombinations}
\title{Obtain all possible marker combinations}
\usage{
GetMarkerCombinations(og)
}
\arguments{
\item{og}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}
}
\value{
(dataFrame) Returns a dataFrame containing all possible marker combinations
}
\description{
Obtain all possible marker combinations
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "CompensAID"))

# Get combinations
mc <- GetMarkerCombinations(og = file)

}
\seealso{
\code{\link{CompensAID}}
}
