% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPopulations.R
\name{GetPopulations}
\alias{GetPopulations}
\title{Obtain (segmented) populations}
\usage{
GetPopulations(og, primary, secondary, co.input, sd.input)
}
\arguments{
\item{og}{(FlowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}

\item{primary}{(character): Name of the primary marker.}

\item{secondary}{(character): Name of the secondary marker.}

\item{co.input}{(dataFrame) dataFrame with the output of the density-based cut-off detection.}

\item{sd.input}{(numerical): Numerical value determining the distance between the primary negative and positive population.}
}
\value{
(list) Returns a list with the negative population of the primary and secondary marker, and positive population of the primary marker.
}
\description{
Obtain (segmented) populations
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "250410.fcs", package = "compensAID"))

# Marker names
primary.marker <- "CD19"
secondary.marker <- "IgL"

# Density-based cut-off detection
center.plot <- 2
co <- DensityGating(og = file,
                    cp.value = center.plot)

# Parameter for the distance between the primary positive and negative population
separation.distance <- 0.25

# Obtain populations
pop <- GetPopulations(og = file,
                      primary = primary.marker,
                      secondary = secondary.marker,
                      co.input = co,
                      sd.input = separation.distance)
}
\seealso{
\code{\link{CompensAID}}, \code{\link{DensityGating}}
}
