% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSegment.R
\name{GetSegment}
\alias{GetSegment}
\title{Calculate segment information}
\usage{
GetSegment(
  population,
  primary.channel,
  secondary.channel,
  segment,
  range.input
)
}
\arguments{
\item{population}{(matrix): Matrix of the population that is segmented.}

\item{primary.channel}{(character): Name of the primary channel.}

\item{secondary.channel}{(character): Name of the secondary channel.}

\item{segment}{(numerical) Segment number for which the information is obtained.}

\item{range.input}{(numerical): Numerical value defining the width of each segment.}
}
\value{
(dataFrame) Returns a dataframe with the minimum segment value, maximum segment value, events within the segment, and mean fluorescence intensity.
}
\description{
Calculate segment information
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "compensAID"))

# Marker names
primary.marker <- "CD19"
secondary.marker <- "CD3"

# Channel names
cp <- "PE-Cy5-A"
cs <- "PE-Cy7-A"

# Density-based cut-off detection
center.plot <- 2
co <- DensityGating(og = file,
                    cp.value = center.plot)

# Parameter for the distance between the primary positive and negative population
separation.distance <- 0.25

# Obtain populations
pop <- GetPopulations(og = file,
                      primary = primary.marker,
                      secondary = secondary.marker,
                      co.input = co,
                      sd.input = separation.distance)

# Select primary positive population
population.positive <- pop$primary.positive

# Get the events that fall within the fourth segment
segment.value <- 4
range.value <- (max(population.positive[, cp]) - min(population.positive[, cp]))/segment.value
info <- GetSegment(population = population.positive,
                   primary.channel = cp,
                   secondary.channel = cs,
                   segment = segment.value,
                   range.input = range.value)
}
\seealso{
\code{\link{CompensAID}}, \code{\link{DensityGating}}, \code{\link{UpdateMatrixInfo}}, \code{\link{GetPopulations}}
}
