% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDotSSI.R
\name{PlotDotSSI}
\alias{PlotDotSSI}
\title{Plot dot plot of Secondary Stain Index scores}
\usage{
PlotDotSSI(output.compensAID, og, primary, secondary, showScores = FALSE)
}
\arguments{
\item{output.compensAID}{(matrix): Matrix containing the SSI output.}

\item{og}{(flowFrame): FlowFrame containing the expression matrix, channel names, and marker names.}

\item{primary}{(character): Minimum SSI value shown in matrix.}

\item{secondary}{(character): Maximum SSI value shown in matrix.}

\item{showScores}{(boolean): Boolean variable determining if scores and gating should be visualized.}
}
\value{
(ggplot2) Returns figure containing a dot plot of the compensAID output.
}
\description{
Plot dot plot of Secondary Stain Index scores
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "CompensAID"))

# Run compensAID tool
segments <- 4
min.events <- 20
center <- 2
distance.populations <- 0.25
compensAID.res <- CompensAID(ff = file,
                             range.value = segments,
                             events.value = min.events,
                             center.plot = center,
                             separation.distance = distance.populations)

# Marker names
primary.marker <- "CD19"
secondary.marker <- "CD3"

# Plot matrix
figure <- PlotDotSSI(output.compensAID = compensAID.res,
                     og = file,
                     primary = primary.marker,
                     secondary = secondary.marker,
                     showScores = TRUE)

}
\seealso{
\code{\link{CompensAID}}
}
