% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMatrix.R
\name{PlotMatrix}
\alias{PlotMatrix}
\title{Plot Secondary Stain Index matrix}
\usage{
PlotMatrix(output.compensAID, sample, limit.min = -3, limit.max = 3)
}
\arguments{
\item{output.compensAID}{(matrix): Matrix containing the SSI output.}

\item{sample}{(character): Name of sample.}

\item{limit.min}{(numerical): Minimum SSI value shown in matrix.}

\item{limit.max}{(numerical): Maximum SSI value shown in matrix.}
}
\value{
(ggplot2) Returns figure containing a matrix of the total compensAID output.
}
\description{
Plot Secondary Stain Index matrix
}
\examples{
# Import FCS file
file <- flowCore::read.FCS(system.file("extdata", "68983.fcs", package = "compensAID"))

# Run compensAID tool
segments <- 4
min.events <- 20
center <- 2
distance.populations <- 0.25
compensAID.res <- CompensAID(ff = file,
                             range.value = segments,
                             events.value = min.events,
                             center.plot = center,
                             separation.distance = distance.populations)

# Plot matrix
figure <- PlotMatrix(output.compensAID = compensAID.res,
                     sample = "68983.fcs")

}
\seealso{
\code{\link{CompensAID}}
}
