% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithinLimit.R
\name{WithinLimit}
\alias{WithinLimit}
\title{Calculate Secondary Stain Index}
\usage{
WithinLimit(
  population,
  og,
  primary,
  secondary,
  min = 10,
  max = 90,
  si.input,
  sd.input,
  co.input
)
}
\arguments{
\item{si.input}{(dataFrame): DataFrame containing the SSI info.}

\item{primary.channel}{(character): Name of the primary channel.}

\item{secondary.channel}{(character): Name of the secondary channel.}

\item{segment}{(numerical): Numerical value determining for which segment the SSI should be calculated.}
}
\value{
(numerical) Returns the Secondary Stain Index score.
}
\description{
Calculate Secondary Stain Index
}
\examples{
# Import FCS file
file <- flowCore::read.FCS("path/to/exampleFCS.fcs")

# Obtain all possible marker combinations
mc <- GetMarkerCombinations(og = file)

# Density-based cut-off detection
center.plot <- 2
co <- DensityGating(og = file,
                    cp.value = center.plot)

# Empty SSI(-info) matrix/dataFrame
range.value <- 4
separation.distance <- 0.25
sm <- EmptyMatrix(og = file)
si <- EmptyMatrixInfo(og = file,
                      rv.input = range.value,
                      mc.input = mc,
                      co.input = co,
                      sd.input = separation.distance)

# Gating positive and negative population
primary.marker <- "CD19"
secondary.marker <- "IgL"
pop <- GetPopulations(og = file,
                      primary = primary.marker,
                      secondary = secondary.marker,
                      co.input = co,
                      sd.input = separation.distance)

# Adjust limits
pop.limit <- WithinLimit(population = pop,
                         og = file,
                         primary = primary.marker,
                         secondary = secondary.marker,
                         min = 10,
                         max = 90,
                         si.input = si,
                         sd.input = separation.distance,
                         co.input = co)

}
\seealso{
\code{\link{CompensAID}}, \code{\link{EmptyMatrixInfo}}, \code{\link{DensityGating}}
}
